Added AI Parsing to Python and Java
I mean it works.
The UI is pretty bad, and honestly idk what to do.
To all those reading this, thank you.
I would appreciate any pull requests to fix the ui of the ai.
My school is starting soon so I won't be doing project anymore.
Thank you for considering looking at this project.
It was fun..